Fix for command overrun
Due to a flaw in my design, the user was allowed to queue up an unlimited
number of commands by simply hitting the buttons quickly; large numbers of
commands can get queued up quickly, finally leaving Pierogi in a confused
state.  This fix should only allow a single command to be processed at a
time, ignoring any input from the user while it is in progress.