Still fixing install, more keyset work
I'm still trying to get install to do what I want, but it still either
doesn't remove pierogi.sudoers when it should, or does remove it when it
shouldn't.  For now, I'll have to ask users to clean up that file manually
on uninstall. :(  In other news, some cleanup work on the keysets, along
with a few new keysets (Denon, Hitachi, Pinnacle).