A couple of UI additions
This update includes support for searching the Keyset and Device lists using
the keyboard.  Also, a new preferences item added to choose a "default keyset";
currently, the only thing the default can do is override the volume controls.
A Digital Stream keyset bug was fixed, a new Grundig satellite keyset and
Daewoo TV keyset added, and a new keyset created for Lexuz.