New Keysets
I've found a new source of keyset data, so I'm now able to add a few
more keysets for brands that were not previously supported.  A first
pass has been made at keysets for Crown Mustang, Hantarex, Keymat,
Konka, and Wiwa, and new keysets added to ADB (I-Can), Sony, and
Triax.  Also, some tweaks to the macro UI, fixes to a Cisco keyset
and an LG timing issue, and improvements to thread management when
the app exits.