Keyset Memory Management Bugfix
This update fixes a serious bug in the handling of keysets: keyset
protocol data must not be deleted while a command is still in progress.
Also in this update, a first pass at keysets for Changhong, Frontech,
and Sinotec; a new Thomson keyset; and fixes for an ADB (I-Can) keyset.