Bugfix Time
This commit mainly includes a fix for an error in storing the make name
for favorite keysets, causing those keysets to be forgotten.  Also, pushed
most "setValue()" calls to destructors, so that QSettings does not sync()
to the flash storage.  (Need to finish pushing the rest of the setValue()
calls to the destructors asap.)  And, first pass at keysets for Comag,
OpenBox, SilverCrest, and Xoro, and new projector keysets for Sanyo.