Licensing Update
I'm finally trying to bring this code into line with the guidelines for
using the GPL, and so I've started adding the standard GPL boilerplate
to the top of every file, as well as including a "COPYING" file.  I'll
probably need to continue to massage the files to make them compliant.
Also in this commit, some UI changes -- "Advanced Settings" has been
moved into the preferences window, and the Air Conditioner panels have
been split into two different tab sets, one for stateless AC and one for
stateful AC.  Also, a first pass at keysets for General Electric,
Majestic, and Minolta, and more keysets for Emerson, Insignia, and
Strong.