Extras Candidate
This update finally brings the documentation up to the current status,
more or less.  I still believe more cleanup would be useful, but
pushing a modern version of Pierogi into Extras should probably take
precedence.  Also in this update, a new keyset for Grundig (which is
ultimately just a link to Philips TV Keyset 1).